tests: use pytest_asyncio in "auto" mode
Use pytest_asyncio in auto mode, instead of "strict", as tests fail with
newer versions.

It was either this, or annotating our async fixtures with
pytest_asyncio.fixture.  That's not too difficult, but pytest_asyncio
says: "This mode requires the simplest tests and fixtures configuration
and is recommended for default usage unless the same project and its
test suite should execute tests from different async frameworks, e.g.
asyncio and trio", which is not the case here.