types: convert an str/bytes/bytearray to str
In PrometheusServer, self.server_address[0]'s type is
  Union[str, bytes, bytearray]

server_address is set by a socket.getsockname() call, through:
  http.server.ThreadingHTTPServer.__init__()
    -> http.server.HTTPServer.__init__()
    -> socketserver.TCPServer.__init__()

mypy complained, since we passed it to HTTPConnection in
test_prometheus.py, which expects an str.

Pass it through str() in the constructor, as that's probably better for
other purposes (like e.g. log output)