types: fix a SystemExit mismatch in test_argparse_misc
SystemExit.code can be None, because sys.exit() can take None (== 0).

SystemExit.code can also be str() in the case where one wants to raise
SystemExit with a reason.

We were comparing this with an integer (== 0 and < 0) in
test_argparse_misc, which rightfully made mypy unhappy. Handle these
cases through an inline if and int().