tests: remove support for CSafeLoader
This is triggered by typing errors, where it's proving to be just a
headache. It can be resolved, for example with:
  loader = cast(Type[SafeLoader), CSafeLoader)

However, a) this was a premature optimization to begin with, with the
YAML file being a static fixture, and with speed benefits being
non-existent on my computer b) this creates effectively two different
code paths, in what is supposed to be a test where we'd like to keep
differences to a minimum.

Thus remove the C code path entirely and just use SafeLoader.